# lib/api-zod/src/generated/types/dayStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Calorie Tracker API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DayStats {
  date: Date;
  totalCalories: number;
  goalCalories: number;
}